victara: proprietary: Use our old libqc-opt.so for vendor perf ext
When the blobs were updated for M, libqti-perfd-client.so appeared
to have replaced libqc-opt.so and we made the corresponding change
to our proprietary blobs.

However, that wasn't really the case.  The extension lib
libqti-perfd-client.so is paired with perfd and not with mpdecision and
we are still using mpdecision.  As far as I can tell, motorola simply
stopped shipping libqc-opt.so because their power HAL doesn't use it.

Our power HAL does use the extension lib so let's use it to talk to
mpdecision and do something more reasonable with the performance hints.

Change-Id: I6770af55d9c7303f16573231840ef0e26bc555eb